Migrate to DTO and in the change delete logic

+++ src/main/java/com/example/obrestdata_jpa/Services/AuthorService.java
@@ -1,10 +1,13 @@
 package com.example.obrestdata_jpa.Services;
 
 import ch.qos.logback.classic.Logger;
+import com.example.obrestdata_jpa.DTO.AuthorDTO;
 import com.example.obrestdata_jpa.Entities.Author;
+import com.example.obrestdata_jpa.Entities.Book;
 import com.example.obrestdata_jpa.Error.BadRequestException;
 import com.example.obrestdata_jpa.Error.NotFoundException;
 import com.example.obrestdata_jpa.Repositories.AuthorRepository;
+import com.example.obrestdata_jpa.Repositories.BookRepository;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
@@ -13,26 +16,38 @@ import java.util.List;
 
 @Service
 public class AuthorService {
+
     private final AuthorRepository authorRepository;
+    private final BookRepository bookRepository;
     private final Logger log = (Logger) LoggerFactory.getLogger(BookService.class);
 
-    public AuthorService(AuthorRepository authorRepository) {
+    public AuthorService(AuthorRepository authorRepository, BookRepository bookRepository) {
         this.authorRepository = authorRepository;
+        this.bookRepository = bookRepository;
     }
+
 
     /*
      * @return List<Author>
      */
-    public List<Author> findAll() {
-        return this.authorRepository.findAll();
+    public List<AuthorDTO> findAll() {
+        List<AuthorDTO> authorDTOList = this.authorRepository.findAll().stream().map(AuthorDTO::new).toList();
+        if (authorDTOList.isEmpty()) {
+            throw new NotFoundException("Author not found");
+        }
+        return authorDTOList;
     }
 
     /*
      * @params id
      * @return Author or Null
      */
-    public Author findById(Long id) {
-        return this.authorRepository.findById(id).orElse(null);
+    public AuthorDTO findById(Long id) {
+        Author author = this.authorRepository.findById(id).orElse(null);
+        if (author == null) {
+            throw new NotFoundException("Author not found");
+        }
+        return new AuthorDTO(author);
     }
 
     /*
@@ -100,6 +115,21 @@ public class AuthorService {
      * @return Void
      */
     public void deleteAll() {
+        List<Author> authorList = this.authorRepository.findAll();
+
+        if (authorList.isEmpty()) {
+            throw new NotFoundException("Not found author to delete");
+        }
+        /* Remove all books from the author, to avoid the error of foreign key constraint: "Cannot delete or update a parent row: a foreign key constraint fails"
+         *  SET NULL the author of de book and remove the list of books from the author after that delete all authors
+         */
+        authorList.forEach(author -> {
+            List<Book> auhtorBookList = this.bookRepository.findBookByAuthorId(author.getId());
+            author.removeBook(auhtorBookList);
+        });
+
         this.authorRepository.deleteAll();
+
+
     }
 }
